Add swiping left/right through dining menus
